Add factoring of files
