+++ packages/invoices/src/InboundShipment/DetailView/modals/InboundLineEdit/index.ts
@@ -1,0 +1,1 @@
+export * from './InboundLineEdit';
